Allow GHC to dump primop modules
This patch adds two commands to GHC:
- --print-prim-module: prints the contents of GHC.Internal.Prim
- --print-prim-wrappers-module: prints the contents of GHC.Internal.PrimopWrappers

These two commands can be used in ghc-internal's Setup.hs to generate
these modules. This ensures that ghc-internal's primop code is always in
sync with the GHC that builds it. It also avoids having to share code
(primops.txt.pp and friends) between ghc-internal and ghc.