localRegistersConflict: account for assignment LHS
This commit fixes a serious oversight in GHC.Cmm.Sink.conflicts,
specifically the code that computes which local registers conflict
between an assignment and a Cmm statement.

If we have:

  assignment: <local_reg> = <expr>
        node: <local_reg> = <other_expr>

then clearly the two conflict, because we cannot move one statement past
the other, as they assign two different values to the same local
register. (Recall that 'conflicts (local_reg,expr) node' is False if and
only if the assignment 'local_reg = expr' can be safely commuted past
the statement 'node'.)

The fix is to update 'GHC.Cmm.Sink.localRegistersConflict' to take into
account the following two situations:

  (1) 'node' defines the LHS local register of the assignment,
  (2) 'node' defines a local register used in the RHS of the assignment.

The bug is precisely that we were previously missing condition (1).

Fixes #26550